refactor: move courseItem to courseList folder
