+++ frontend/src/components/adminPage/courseList/courseItem/courseItem.module.scss
