+++ frontend/src/components/adminPage/courseList/courseItem/courseItem.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { deleteCourse } from '../../../features/course/courseSlice';
+import { deleteCourse } from '../../../../features/course/courseSlice';
 import styles from './courseItem.module.scss';
-import EditCourseModal from '../editCourse/editCourseModal';
+import EditCourseModal from '../../editCourse/editCourseModal';
 
 interface Course {
   id: string;
